refact : domain 객체 관련, field 중 double로 잘못 설정되어 있던 부분을 BigDecimal로 고침

+++ src/main/java/com/billit/investment/service/SettlementService.java
@@ -22,7 +22,7 @@ public class SettlementService {
         Settlement settlement = settlementRepository.findByInvestmentId(investmentId)
                 .orElseThrow(() -> new IllegalArgumentException("Settlement not found for investmentId: " + investmentId));
 
-        BigDecimal settlementAmount = repaymentAmount.multiply(BigDecimal.valueOf(settlement.getSettlementRatio()));
+        BigDecimal settlementAmount = repaymentAmount.multiply(settlement.getSettlementRatio());
         BigDecimal principal = calculatePrincipal(settlementAmount);
         BigDecimal profit = settlementAmount.subtract(principal);
